separate combat and rest sites
